use-latest-values - move to yes-no-unknown.ts
