+++ src/app/user/registration/_components/donor-form.tsx
@@ -17,7 +17,7 @@ import { BarkFormOption } from "@/components/bark/bark-form";
 import { postRegistrationRequest } from "@/app/user/registration/_actions/post-registration-request";
 import { RegistrationRequest } from "@/lib/services/registration";
 import { DogAntigenPresence } from "@/lib/data/db-enums";
-import { YesNoUnknown } from "@/lib/bark/enums/yes-no";
+import { YesNoUnknown } from "@/lib/bark/enums/yes-no-unknown";
 import { DogGender } from "@/lib/bark/models/dog-gender";
 import { UserResidency } from "@/lib/data/db-enums";
 import { BARK_UTC } from "@/lib/utilities/bark-time";
