+++ src/lib/admin/admin-models.ts
@@ -1,4 +1,4 @@
-import { YesNoUnknown } from "../bark/enums/yes-no";
+import { YesNoUnknown } from "../bark/enums/yes-no-unknown";
 import { DogGender } from "../bark/models/dog-gender";
 
 export type IncompleteProfile = {
