+++ src/lib/data/db-models.ts
@@ -15,7 +15,7 @@
 //
 
 import { DogAntigenPresence } from "./db-enums";
-import { YesNoUnknown } from "../bark/enums/yes-no";
+import { YesNoUnknown } from "../bark/enums/yes-no-unknown";
 import { DogGender } from "../bark/models/dog-gender";
 import { UserResidency } from "./db-enums";
 import { CallOutcome, PosNegNil, ReportedIneligibility } from "./db-enums";
