+++ tests/_fixtures.ts
@@ -21,7 +21,7 @@ import {
   DbCallGen,
 } from "@/lib/data/db-models";
 import { DOG_ANTIGEN_PRESENCE, DogAntigenPresence } from "@/lib/data/db-enums";
-import { YES_NO_UNKNOWN, YesNoUnknown } from "@/lib/bark/enums/yes-no";
+import { YES_NO_UNKNOWN, YesNoUnknown } from "@/lib/bark/enums/yes-no-unknown";
 import { DOG_GENDER, DogGender } from "@/lib/bark/models/dog-gender";
 import { USER_RESIDENCY } from "@/lib/data/db-enums";
 import { dbInsertAdmin, dbSelectAdmin } from "@/lib/data/db-admins";
