+++ tests/admin/get-num-incomplete-profiles.test.ts
@@ -8,7 +8,7 @@ import {
 } from "../_fixtures";
 import { Pool } from "pg";
 import { PARTICIPATION_STATUS } from "@/lib/data/db-enums";
-import { YES_NO_UNKNOWN } from "@/lib/bark/enums/yes-no";
+import { YES_NO_UNKNOWN } from "@/lib/bark/enums/yes-no-unknown";
 import { DOG_GENDER } from "@/lib/bark/models/dog-gender";
 import {
   DEFAULT_DATE_TIME_FORMAT,
